okay now:

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	unsigned int count;
+	size_t count;
 
 	count = 0;
 	while (h != NULL)
@@ -21,7 +21,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%u] %s\n", (unsigned int)count, h->str);
+			printf("[%lu] %s\n", count, h->str);
 		}
 		h = h->next;
 		count++;
